CDSID-1168 Fix review comments
