+++ README.md
@@ -2,11 +2,11 @@
 # cds-imports-dds
 
 ### End Points:
-POST /declarations with header X-EORI-Number 
+POST /declarations with header `X-EORI-Identifier` 
 
 #### Examples:
 ```bash
-curl -d '{"lrn": "1234"}' -H "X-EORI-Number: eori123" -H "Content-Type: application/json" -XPOST http://localhost:9759/declarations
+curl -d '{"lrn": "1234"}' -H "X-EORI-Identifier: eori123" -H "Content-Type: application/json" -XPOST http://localhost:9759/declarations
 ```
 
 ### License
